requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
email-validator==2.1.0.post1
Flask-Admin==1.6.1
Flask-Babel==4.0.0
opencv-python==4.9.0.80
piexif==1.1.3
feedgen==0.9.0
celery==5.3.6
redis==5.0.1 